requirements.txt
streamlit
mysql.connector
csv
pandas
plotly.express
streamlit_lottie 
json
dotenv 
os
google.generativeai 
re